Update middleware on racks
