+++ src/middleware/src/storage/index.js
@@ -1,13 +1,12 @@
-import KinveyStorage from 'kinvey-node-sdk/dist/request/src/middleware/src/storage';
-import { isDefined } from 'kinvey-node-sdk/dist/utils';
+import { Storage as NodeStorage, isDefined } from 'kinvey-node-sdk/dist/export';
 import TitaniumDB from './src/titaniumdb';
 
-export default class Storage extends KinveyStorage {
-  getAdapter() {
-    return TitaniumDB.loadAdapter(this.name)
+export default class Storage extends NodeStorage {
+  loadAdapter() {
+    return TitaniumDB.load(this.name)
       .then((adapter) => {
         if (isDefined(adapter) === false) {
-          return super.getAdapter();
+          return super.loadAdapter();
         }
 
         return adapter;
