--- src/middleware/src/storage/src/errors/index.js
@@ -1,6 +1,0 @@
-import NotFoundError from './src/notfound';
-
-// Export
-export {
-  NotFoundError
-};
