--- src/middleware/src/storage/src/errors/src/notfound.js
@@ -1,3 +1,0 @@
-import ExtendableError from 'es6-error';
-
-export default class NotFoundError extends ExtendableError { }
